Test horiztonal resolution attribute
